@@ -1,3 +1,5 @@
-const semver = require('semver');
+import semver from "semver";
 
-module.exports = (channel) => (channel ? (semver.validRange(channel) ? `release-${channel}` : channel) : 'latest');
+export default function (channel) {
+  return channel ? (semver.validRange(channel) ? `release-${channel}` : channel) : "latest";
+}
