@@ -1,7 +1,8 @@
-const SemanticReleaseError = require('@semantic-release/error');
-const ERROR_DEFINITIONS = require('./definitions/errors');
+import SemanticReleaseError from "@semantic-release/error";
+import * as ERROR_DEFINITIONS from "./definitions/errors.js";
 
-module.exports = (code, ctx = {}) => {
-  const {message, details} = ERROR_DEFINITIONS[code](ctx);
+export default function (code, ctx = {}) {
+  const { message, details } = ERROR_DEFINITIONS[code](ctx);
+
   return new SemanticReleaseError(message, code, details);
-};
+}
