@@ -1,22 +1,22 @@
-const path = require('path');
-const readPkg = require('read-pkg');
-const AggregateError = require('aggregate-error');
-const getError = require('./get-error');
+import path from "path";
+import { readPackage } from "read-pkg";
+import AggregateError from "aggregate-error";
+import getError from "./get-error.js";
 
-module.exports = async ({pkgRoot}, {cwd}) => {
+export default async function ({ pkgRoot }, { cwd }) {
   try {
-    const pkg = await readPkg({cwd: pkgRoot ? path.resolve(cwd, String(pkgRoot)) : cwd});
+    const pkg = await readPackage({ cwd: pkgRoot ? path.resolve(cwd, String(pkgRoot)) : cwd });
 
     if (!pkg.name) {
-      throw getError('ENOPKGNAME');
+      throw getError("ENOPKGNAME");
     }
 
     return pkg;
   } catch (error) {
-    if (error.code === 'ENOENT') {
-      throw new AggregateError([getError('ENOPKG')]);
+    if (error.code === "ENOENT") {
+      throw new AggregateError([getError("ENOPKG")]);
     }
 
     throw new AggregateError([error]);
   }
-};
+}
