@@ -1,15 +1,18 @@
-const path = require('path');
-const rc = require('rc');
-const getRegistryUrl = require('registry-auth-token/registry-url');
+import path from "path";
+import rc from "rc";
+import getRegistryUrl from "registry-auth-token/registry-url.js";
 
-module.exports = ({publishConfig: {registry} = {}, name}, {cwd, env}) =>
-  registry ||
-  env.NPM_CONFIG_REGISTRY ||
-  getRegistryUrl(
-    name.split('/')[0],
-    rc(
-      'npm',
-      {registry: 'https://registry.npmjs.org/'},
-      {config: env.NPM_CONFIG_USERCONFIG || path.resolve(cwd, '.npmrc')}
+export default function ({ publishConfig: { registry } = {}, name }, { cwd, env }) {
+  return (
+    registry ||
+    env.NPM_CONFIG_REGISTRY ||
+    getRegistryUrl(
+      name.split("/")[0],
+      rc(
+        "npm",
+        { registry: "https://registry.npmjs.org/" },
+        { config: env.NPM_CONFIG_USERCONFIG || path.resolve(cwd, ".npmrc") }
+      )
     )
   );
+}
