@@ -1,15 +1,17 @@
-const normalizeUrl = require('normalize-url');
+import normalizeUrl from "normalize-url";
 
-module.exports = (
-  {name},
-  {env: {DEFAULT_NPM_REGISTRY = 'https://registry.npmjs.org/'}, nextRelease: {version}},
+export default function (
+  { name },
+  { env: { DEFAULT_NPM_REGISTRY = "https://registry.npmjs.org/" }, nextRelease: { version } },
   distTag,
   registry
-) => ({
-  name: `npm package (@${distTag} dist-tag)`,
-  url:
-    normalizeUrl(registry) === normalizeUrl(DEFAULT_NPM_REGISTRY)
-      ? `https://www.npmjs.com/package/${name}/v/${version}`
-      : undefined,
-  channel: distTag,
-});
+) {
+  return {
+    name: `npm package (@${distTag} dist-tag)`,
+    url:
+      normalizeUrl(registry) === normalizeUrl(DEFAULT_NPM_REGISTRY)
+        ? `https://www.npmjs.com/package/${name}/v/${version}`
+        : undefined,
+    channel: distTag,
+  };
+}
