@@ -1,16 +1,16 @@
-const path = require('path');
-const execa = require('execa');
-const getRegistry = require('./get-registry');
-const getChannel = require('./get-channel');
-const getReleaseInfo = require('./get-release-info');
+import path from "path";
+import execa from "execa";
+import getRegistry from "./get-registry.js";
+import getChannel from "./get-channel.js";
+import getReleaseInfo from "./get-release-info.js";
 
-module.exports = async (npmrc, {npmPublish, pkgRoot}, pkg, context) => {
+export default async function (npmrc, { npmPublish, pkgRoot }, pkg, context) {
   const {
     cwd,
     env,
     stdout,
     stderr,
-    nextRelease: {version, channel},
+    nextRelease: { version, channel },
     logger,
   } = context;
 
@@ -21,12 +21,12 @@ module.exports = async (npmrc, {npmPublish, pkgRoot}, pkg, context) => {
 
     logger.log(`Publishing version ${version} to npm registry on dist-tag ${distTag}`);
     const result = execa(
-      'npm',
-      ['publish', basePath, '--userconfig', npmrc, '--tag', distTag, '--registry', registry],
-      {cwd, env, preferLocal: true}
+      "npm",
+      ["publish", basePath, "--userconfig", npmrc, "--tag", distTag, "--registry", registry],
+      { cwd, env, preferLocal: true }
     );
-    result.stdout.pipe(stdout, {end: false});
-    result.stderr.pipe(stderr, {end: false});
+    result.stdout.pipe(stdout, { end: false });
+    result.stderr.pipe(stderr, { end: false });
     await result;
 
     logger.log(`Published ${pkg.name}@${version} to dist-tag @${distTag} on ${registry}`);
@@ -35,10 +35,10 @@ module.exports = async (npmrc, {npmPublish, pkgRoot}, pkg, context) => {
   }
 
   logger.log(
-    `Skip publishing to npm registry as ${npmPublish === false ? 'npmPublish' : "package.json's private property"} is ${
+    `Skip publishing to npm registry as ${npmPublish === false ? "npmPublish" : "package.json's private property"} is ${
       npmPublish !== false
     }`
   );
 
   return false;
-};
+}
