@@ -1,5 +1,5 @@
-const {isString, isNil, isBoolean} = require('lodash');
-const getError = require('./get-error');
+import { isBoolean, isNil, isString } from "lodash-es";
+import getError from "./get-error.js";
 
 const isNonEmptyString = (value) => isString(value) && value.trim();
 
@@ -9,14 +9,14 @@ const VALIDATORS = {
   pkgRoot: isNonEmptyString,
 };
 
-module.exports = ({npmPublish, tarballDir, pkgRoot}) => {
-  const errors = Object.entries({npmPublish, tarballDir, pkgRoot}).reduce(
+export default function ({ npmPublish, tarballDir, pkgRoot }) {
+  const errors = Object.entries({ npmPublish, tarballDir, pkgRoot }).reduce(
     (errors, [option, value]) =>
       !isNil(value) && !VALIDATORS[option](value)
-        ? [...errors, getError(`EINVALID${option.toUpperCase()}`, {[option]: value})]
+        ? [...errors, getError(`EINVALID${option.toUpperCase()}`, { [option]: value })]
         : errors,
     []
   );
 
   return errors;
-};
+}
